moved Hypergrid.js to hypergrid/index.js, etc.
