+++ src/hypergrid/selection.js
